pyproject.toml
[project]
name = "easyminer-backend"
version = "0.1.0"
description = "Easyminer backend"
authors = [{ name = "Radim Sückr", email = "sucr02@vse.cz" }]
dependencies = [
	"alembic>=1.13.2",
	"cachetools>=6.2.1",
	"celery[redis,tblib]>=5.4.0",
	"cleverminer>=1.2.2",
	"defusedxml>=0.7.1",
	"efficient-apriori>=2.0.6",
	"fastapi[standard]>=0.111.1",
	"pandas>=2.2.3",
	# "pyARC @ file:///Users/radimsuckr/Work/pyARC",
	"pyARC @ https://github.com/radimsuckr/pyARC.git",
	"pydantic-settings>=2.4.0",
	"pydantic-xml[lxml]>=2.17.0",
	"pyhumps>=3.8.0",
	"python-multipart>=0.0.20",
	"sqlalchemy[asyncio,pymysql,aiomysql]>=2.0.31",
]
readme = "readme.md"
requires-python = ">= 3.13"

[dependency-groups]
dev = [
	"celery-types>=0.23.0",
	"lxml-stubs>=0.5.1",
	"mypy>=1.11.0",
	"pandas-stubs>=2.2.3.250308",
	"pytest>=8.3.1",
	"pytest-asyncio>=0.25.3",
	"pytest-cov>=6.0.0",
	"ruff>=0.9.10",
	"types-cachetools>=6.2.0.20251022",
]

[tool.ruff]
target-version = "py313"
line-length = 120

[tool.ruff.format]
line-ending = "lf"

[tool.ruff.lint]
extend-select = ["E4", "E7", "E9", "F", "FAST", "I", "INP", "UP"]

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"
